+++ CHANGES.md
@@ -4,5 +4,6 @@
 
 -   Add Sort Button #30
 -   Add props to displayName override function #37
+-   Add sortable prop to schema #38
 
 ### Unreleased